+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/typeSystem/TSDeploymentTableMustExistForItemExtendingGenericItem.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -50,7 +50,7 @@ class TSDeploymentTableMustExistForItemExtendingGenericItem : AbstractTSInspecti
         severity: HighlightSeverity
     ) {
         // skip non-ComposedType
-        if (HybrisConstants.TS_META_VIEW_TYPE.equals(dom.metaType.stringValue, true) || HybrisConstants.TS_META_COMPOSED_TYPE.equals(dom.metaType.stringValue, true)) return
+        if (HybrisConstants.TS_META_VIEW_TYPE.equals(dom.metaType.stringValue, true) || HybrisConstants.TS_COMPOSED_TYPE.equals(dom.metaType.stringValue, true)) return
 
         val itemTypeCode = dom.code.stringValue ?: return