+++ src/com/intellij/idea/plugin/hybris/system/type/meta/model/TSMetaItem.kt
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -61,7 +61,7 @@ interface TSMetaItem : TSMetaClassifier<ItemType> {
         val type: String?
         val isSelectionOf: String?
         val metaType: String
-            get() = HybrisConstants.TS_META_TYPE_ATTRIBUTE_DESCRIPTOR
+            get() = HybrisConstants.TS_TYPE_ATTRIBUTE_DESCRIPTOR
         val isDeprecated: Boolean
         val isAutoCreate: Boolean
         val isRedeclare: Boolean